refactor: reduce duplication via new `TrustDomainStore` trait
Reduce duplication between `SpiffeClientConfigStream` and
`SpiffeServerConfigStream` by breaking out `build_root_store` to
`TrustDomainStore` trait to permit reuse.
